Substitution.c
A substitution cipher takes each character in a message and replaces it with a different character according to fixed rules. Every occurrence of one character will be substituted with the same replacement character. In this C language code, the user provides an alphabetic order of choice to encrypt a plaintext. For example, if I choose my alphabetic order as NQXPOMAFTRHLZGECYJIUWSKDVB, the letter A will become N, letter B will become Q and so on and so forth.
To “decrypt” the message, the receiver of the message would need to know the key, so that they can reverse the process: translating the encrypted text (ciphertext) back into the original message (plaintext)